cms-admin: Split TipTap translation stories into their own file
TipTapRichTextBlock.stories.tsx was becoming unwieldy with the translation
stories mixed in alongside the block's other feature stories. Move them into
TipTapRichTextBlockTranslation.stories.tsx, following the same pattern
already used for TipTapRichTextBlockChildBlocks.stories.tsx.
